+++ src/injectionKeys.ts
@@ -6,3 +6,4 @@ export const CMDK_STATE: InjectionKey<UseCommandMenuReturn> = Symbol('cmdk-state
 export const CMDK_LOADING: InjectionKey<() => boolean> = Symbol('cmdk-loading')
 export const CMDK_CLOSE_ON_SELECT: InjectionKey<() => boolean> = Symbol('cmdk-close-on-select')
 export const CMDK_SELECT_HANDLER: InjectionKey<(item: CommandItemData) => void> = Symbol('cmdk-select-handler')
+export const CMDK_ITEM_INDEX_MAP: InjectionKey<() => Map<string, number>> = Symbol('cmdk-item-index-map')